- Refactored ScreenSwitcher into MainScreen + MainCubit / MainState Made the switching process in separate Cubit & Screen * Provided the HomeCubit to HomeScreens in the screens of the bottom nav bar instead of that provider in routes as there is no navigation * Separated the screens and bottom nav bar items in separate files for Organization
