+++ lib/core/routing/app_router.dart
@@ -1,8 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:mentorship/core/di/dependency_injection.dart';
 import 'package:mentorship/core/routing/routes.dart';
 import 'package:mentorship/features/home/logic/home_cubit.dart';
 import 'package:mentorship/features/home/ui/screens/home_screen.dart';
+import 'package:mentorship/features/main/logic/cubit/main_cubit.dart';
+import 'package:mentorship/features/main/ui/screens/main_screen.dart';
 
 class AppRouter {
   Route? generateRoute(RouteSettings settings) {
@@ -10,12 +13,16 @@ class AppRouter {
     // final arguments = settings.arguments;
 
     switch (settings.name) {
-      case Routes.homeScreen:
+      case Routes.mainScreen:
         return MaterialPageRoute(
           builder: (_) => BlocProvider(
-            create: (context) => HomeCubit(),
-            child: const HomeScreen(),
+            create: (context) => MainCubit(),
+            child: MainScreen(),
           ),
+        );
+      case Routes.homeScreen:
+        return MaterialPageRoute(
+          builder: (_) => const HomeScreen(),
         );
 
       default:
