+++ lib/core/routing/routes.dart
@@ -1,3 +1,4 @@
 class Routes {
+  static const String mainScreen = 'mainScreen';
   static const String homeScreen = 'homeScreen';
 }
