+++ lib/features/home/data/models/product.dart
@@ -1,0 +1,7 @@
+class Product {
+  final String imageUrl;
+  final String name;
+  final double price;
+
+  Product({required this.imageUrl, required this.name, required this.price});
+}
