+++ lib/features/home/ui/screens/home_screen.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:mentorship/features/home/logic/home_cubit.dart';
 import 'package:mentorship/features/home/ui/widgets/banners_section.dart';
 import 'package:mentorship/features/home/ui/widgets/categories_list.dart';
@@ -18,7 +17,7 @@ class HomeScreen extends StatelessWidget {
         child: Column(
           mainAxisAlignment: MainAxisAlignment.start,
           children: [
-            SizedBox(height: 20.h),
+            const SizedBox(height: 20),
             const HomeHeader(),
             const SearchBarWithFilter(),
             const BannersSection(),
