+++ lib/features/main/logic/cubit/main_state.dart
@@ -1,0 +1,3 @@
+abstract class MainState {}
+
+final class MainInitial extends MainState {}
